% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult_ml1_cat.R, R/rdomains.R
\name{adult_ml1_cat}
\alias{adult_ml1_cat}
\title{Probability that Domain Hosts Adult Content Based on features of Domain Name and Suffix alone.}
\usage{
adult_ml1_cat(domains = NULL)

adult_ml1_cat(domains = NULL)
}
\arguments{
\item{domains}{vector of domain names}

\item{domains}{vector of domain names}
}
\value{
data.frame with original list and content category of the domain

data.frame with original list and content category of the domain
}
\description{
Uses a validated ML model that uses keywords in the domain name
and suffix to predict probability that the domain hosts adult content. 
See \url{https://github.com/soodoku/domain_classifier/tree/master/domain_name} for details.

Uses a validated ML model that uses keywords in the domain name
and suffix to predict probability that the domain hosts adult content. 
See \url{https://github.com/soodoku/domain_classifier/tree/master/domain_name} for details.
}
\examples{
adult_ml1_cat(domains="http://www.google.com")

adult_ml1_cat(domains="http://www.google.com")

}

