% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmoz_cat.R
\name{dmoz_cat}
\alias{dmoz_cat}
\title{Get Category from DMOZ}
\usage{
dmoz_cat(domains = NULL, use_file = NULL)
}
\arguments{
\item{domains}{vector of domain names}

\item{use_file}{path to the dmoz file, which can be downloaded using \code{\link{get_dmoz_data}}}
}
\value{
data.frame with original list and content category of the domain
}
\description{
Fetches category (or categories) of content hosted by a domain according to DMOZ.
The function checks if path to the DMOZ file is provided by the user.
If not, it looks for \code{dmoz_domain_cateory.csv} in the working directory. It also returns
results for prominent subdomains.
}
\examples{
\dontrun{
dmoz_cat(domains = "http://www.google.com")
dmoz_cat(domains = c("http://www.google.com", "http://plus.google.com"))
}
}
