% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virustotal_cat.R
\name{domain_report}
\alias{domain_report}
\title{Get Virustotal Category}
\usage{
domain_report(domain = NULL, apikey = NULL)
}
\arguments{
\item{domain}{domain name}

\item{apikey}{virustotal API key}
}
\value{
data frame
}
\description{
Get Virustotal Category
}
\examples{
\dontrun{
domain_report(url="http://www.google.com")
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}

