% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alexa_data.R
\name{get_alexa_data}
\alias{get_alexa_data}
\title{Get Alexa Data}
\usage{
get_alexa_data(outdir = ".", overwrite = FALSE)
}
\arguments{
\item{outdir}{Optional; folder to which you want to save the file; Default is same folder}

\item{overwrite}{Optional; default is FALSE}
}
\description{
Get Alexa Data
}
\examples{
\dontrun{
get_alexa_cat()
}
}
\references{
\url{https://support.alexa.com/hc/en-us/articles/200461990-Can-I-get-a-list-of-top-sites-from-an-API-}
}

