% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glm_shalla}
\alias{glm_shalla}
\title{ML Model}
\format{A list}
\source{
ML model based on shallalist using keywords and domain suffixes,
}
\usage{
glm_shalla
}
\description{
ML Model
}
\author{
Gaurav Sood
}
\keyword{keywords}
\keyword{model}
