% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_data_pkg.R
\name{install_data_package}
\alias{install_data_package}
\title{Install the \code{rdomainsdata} package after checking with the user}
\usage{
install_data_package()
}
\description{
Install the \code{rdomainsdata} package after checking with the user
}
