% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_suffix.R
\name{name_suffix_cat}
\alias{name_suffix_cat}
\title{Get Category Based on Domain Name, Suffix}
\usage{
name_suffix_cat(domains = NULL)
}
\arguments{
\item{domains}{vector of domain names}
}
\value{
data.frame with original list and content category of the domain
}
\description{
Uses a validated ML model that uses keywords in the domain name
and suffix to predict category of content hosted by the domain. 
Currently only supports classification of pornographic content.
}
\examples{
name_suffix_cat(domains="http://www.google.com")

}

