% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_news.R
\name{not_news}
\alias{not_news}
\title{Classify News and Non-News Based on keywords in the URL}
\usage{
not_news(url_list = NULL)
}
\arguments{
\item{url_list}{vector of URLs}
}
\value{
data.frame with 3 columns: url, not_news, news
}
\description{
Based on a slightly amended version of the regular expression used to
classify news, and non-news in:
``Exposure to ideologically diverse news and opinion on Facebook''
by Bakshy, Messing, and Adamic. Science. 2015.
}
\details{
Amendment: sport rather than sports

URL containing any of the following words is classified as soft news:
"sport|entertainment|arts|fashion|style|lifestyle|leisure|celeb|movie|music|gossip|food|travel|horoscope|weather|gadget" # nolint: line_length_linter.

URL containing any of following words is classified as hard news:
"politi|usnews|world|national|state|elect|vote|govern|campaign|war|polic|econ|unemploy|racis|energy|abortion|educa|healthcare|immigration" # nolint: line_length_linter.


Note that it is based on patterns existing in a small set of domains. See paper for details.
}
\examples{
\dontrun{
not_news("http://www.bbc.com/sport")
not_news(c("http://www.bbc.com/sport", "http://www.washingtontimes.com/news/politics/"))
}
}
\references{
\url{https://www.science.org/doi/10.1126/science.aaa1160}
}
