% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shallalist.R
\docType{data}
\name{shalla}
\alias{shalla}
\alias{shallalist}
\title{Shallalist Data}
\format{A data frame with 1,118,363 observations and 2 variables}
\source{
Shallalist,
  \url{http://www.shallalist.de}
}
\usage{
shallalist
}
\description{
A dataset containing 1,118,363 unique domain names and category 
associated with it
}
\author{
Shalla Secure Services KG
}
\keyword{datasets}

