% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult_ml1_cat.R
\name{adult_ml1_cat}
\alias{adult_ml1_cat}
\title{Probability that Domain Hosts Adult Content Based on features of Domain Name and Suffix alone.}
\usage{
adult_ml1_cat(domains = NULL)
}
\arguments{
\item{domains}{required; string; vector of domain names}
}
\value{
data.frame with original list and content category of the domains
}
\description{
Uses a validated ML model that uses keywords in the domain name
and suffix to predict probability that the domain hosts adult content. For
more information see \url{https://github.com/themains/keyword_porn}
}
\examples{
\dontrun{
adult_ml1_cat("http://www.google.com")
}
}
