% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexa_cat.R
\name{alexa_cat}
\alias{alexa_cat}
\title{Get Alexa Category}
\usage{
alexa_cat(domain = NULL, apikey = NULL)
}
\arguments{
\item{domain}{domain name}

\item{apikey}{virustotal API key}
}
\value{
data frame
}
\description{
Get Alexa Category
}
\examples{
\dontrun{
alexa_cat(domain="http://www.google.com")
}
}
\references{
\url{https://docs.aws.amazon.com/AlexaWebInfoService/latest/index.html?ApiReference_UrlInfoAction.html}
}

