% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmoz_cat.R
\name{dmoz_cat}
\alias{dmoz_cat}
\title{Get Category from DMOZ}
\usage{
dmoz_cat(domains = NULL)
}
\arguments{
\item{domains}{vector of domain names}
}
\value{
data.frame with original list and content category of the domain
}
\description{
Fetches category of content hosted by domain from DMOZ.
}
\examples{
dmoz_cat(domains="http://www.google.com")

}

