% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dmoz_data.R
\name{get_dmoz_data}
\alias{get_dmoz_data}
\title{Get DMOZ Data}
\usage{
get_dmoz_data(outdir = ".", overwrite = FALSE)
}
\arguments{
\item{outdir}{Optional; folder to which you want to save the file; Default is same folder}

\item{overwrite}{Optional; default is FALSE. If TRUE, the file is overwritten.}
}
\description{
Downloads, unzips and saves archived version of the DMOZ data. For more details, check:
\url{https://github.com/themains/rdomains/tree/master/data-raw/dmoz/}
}
\examples{
\dontrun{
get_dmoz_data()
}
}
\references{
\url{https://curlie.org/}
}
