% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shalla_data.R
\name{get_shalla_data}
\alias{get_shalla_data}
\title{Get Shalla Data}
\usage{
get_shalla_data(outdir = "./", overwrite = FALSE)
}
\arguments{
\item{outdir}{Optional; folder to which you want to save the file; Default is same folder}

\item{overwrite}{Optional; default is FALSE. If TRUE, the file is overwritten.}
}
\description{
Shallalist service was discontinued in January 2022. This function downloads
the last archived copy (from 1/14/22) that we have preserved on GitHub.
The original service at shallalist.de is no longer available.
Downloads, unzips and saves the final version of shallalist data. By default, saves shalla data
as \code{shalla_domain_category.csv}.
}
\examples{
\dontrun{
get_shalla_data()
}
}
\references{
\url{https://web.archive.org/web/20210502020725/http://www.shallalist.de/}
}
