% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shalla_data.R
\name{get_shalla_data}
\alias{get_shalla_data}
\title{Get Shalla Data}
\usage{
get_shalla_data(outdir = ".", overwrite = FALSE)
}
\arguments{
\item{outdir}{Optional; folder to which you want to save the file; Default is same folder}

\item{overwrite}{Optional; default is FALSE}
}
\description{
Downloads, unzips and saves the latest version of shallalist data.
}
\examples{
\dontrun{
get_shalla_data()
}
}
\references{
\url{http://www.shallalist.de/licence.html}
}

