% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stevenblack_data.R
\name{get_stevenblack_data}
\alias{get_stevenblack_data}
\title{Get Steven Black's Host List Data}
\usage{
get_stevenblack_data(outdir = "./", variant = "base", overwrite = FALSE)
}
\arguments{
\item{outdir}{Optional; folder to save the file in. Default is the
current directory}

\item{variant}{Optional; which variant to download. Options: "base",
"porn", "social", "gambling", "all"}

\item{overwrite}{Optional; default is FALSE. If TRUE, the file is overwritten.}
}
\description{
Downloads the latest version of Steven Black's unified hosts file.
The hosts file contains domains known for serving ads, malware, and tracking.
}
\examples{
\dontrun{
get_stevenblack_data()
get_stevenblack_data(variant = "all")
}
}
\references{
\url{https://github.com/StevenBlack/hosts}
}
