% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shalla.R, R/shallalist.R
\docType{data}
\name{shalla}
\alias{shalla}
\alias{shallalist}
\title{Compare Text}
\format{A data frame with 1,118,363 observations and 2 variables}
\source{
Shallalist,
  \url{http://www.shallalist.de}
}
\usage{
shalla(path_to_ocr = NULL, path_to_gold = NULL, remove_extra_space = TRUE,
  normalize = TRUE)

shallalist
}
\arguments{
\item{path_to_ocr}{path to file containing OCR'd text; required}

\item{path_to_gold}{path to file containing human transcribed text; required}

\item{remove_extra_space}{a dummy indicating whether or not extra spaces should be removed from the OCR file; default is TRUE}

\item{normalize}{add a way to normalize string distance measures -- otherwise longer document means more errors, more distance}
}
\value{
levenshtein distance
}
\description{
This function calculates the edit distance between OCR'd and human transcribed file.
The function by default prints the status of the task you are trying to delete. It will show up as 'deleted' if successful

A dataset containing 1,118,363 unique domain names and category 
associated with it
}
\examples{
\dontrun{
compare_txt(path_to_ocr="path_to_ocr_file", path_to_gold="path_to_gold_file", 
	           remove_extra_space=TRUE)
}
}
\author{
Shalla Secure Services KG
}
\keyword{datasets}

