% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shalla_cat.R
\name{shalla_cat}
\alias{shalla_cat}
\title{Get Category from Shallalist}
\usage{
shalla_cat(domains = NULL, use_new = FALSE)
}
\arguments{
\item{domains}{vector of domain names}

\item{use_new}{use new shallalist database; default is FALSE}
}
\value{
data.frame with original list and content category of the domain
}
\description{
Fetches category of content hosted by domain from Shallalist. If \code{use_new} is TRUE,
\link{\code{get_shall_data}} is invoked, which in turn checks whether the user wants to use a 
local (but newer than what is in the package) shallalist file or download a new one.
}
\examples{
shalla_cat(domains="http://www.google.com")

}

