% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shalla_cat.R
\name{shalla_cat}
\alias{shalla_cat}
\title{Get Category from Shallalist}
\usage{
shalla_cat(domains = NULL, use_file = NULL)
}
\arguments{
\item{domains}{vector of domain names}

\item{use_file}{path to the latest shallalist file downloaded using \code{\link{get_shalla_data}}}
}
\value{
data.frame with original list and content category of the domain
}
\description{
Fetches category of content hosted by domain from Shallalist.
}
\examples{
shalla_cat(domains="http://www.google.com")

}

