% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shalla_cat.R
\name{shalla_cat}
\alias{shalla_cat}
\title{Get Category from Shallalist}
\usage{
shalla_cat(domains = NULL, use_file = NULL)
}
\arguments{
\item{domains}{vector of domain names}

\item{use_file}{path to the latest shallalist file downloaded using \code{\link{get_shalla_data}}}
}
\value{
data.frame with original list and content category of the domain
}
\description{
Fetches category of content hosted by a domain according to Shalla. 
The function checks if path to the shalla file is provided by the user. 
If not, it looks for \code{shalla_domain_cateory.csv} in the working directory.
}
\examples{
\dontrun{
shalla_cat(domains="http://www.google.com")
}
}

