% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trusted_cat.R
\name{trusted_cat}
\alias{trusted_cat}
\title{Get Trusted (McAfee) Category}
\usage{
trusted_cat(domain = NULL, log = FALSE)
}
\arguments{
\item{domain}{domain name (string). Default value is \code{NULL}.}

\item{log}{set whether or not a log file is generated with Selenium. (Boolean). Default value is \code{FALSE}.}
}
\value{
data.frame
}
\description{
There is no API, so the function uses Selenium to open a browser, and the scrape the content.
}
\examples{
\dontrun{
trusted_cat("http://www.google.com")
}
}
\references{
\url{https://www.trustedsource.org/}
}

