% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trusted_cat.R
\name{trusted_cat}
\alias{trusted_cat}
\title{Get Trusted (McAfee) Category}
\usage{
trusted_cat(domain = NULL)
}
\arguments{
\item{domain}{domain name (string)}
}
\value{
data frame
}
\description{
Get Trusted (McAfee) Category
}
\examples{
\dontrun{
trusted_cat(domain="http://www.google.com")
}
}
\references{
\url{https://www.trustedsource.org/}
}

