% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni_cat.R
\name{uni_cat}
\alias{uni_cat}
\title{Get Category from University Domain List}
\usage{
uni_cat(domains = NULL)
}
\arguments{
\item{domains}{vector of domain names}
}
\value{
data.frame with original list and all the other columns from the university json
}
\description{
Fetches university domain json from: 
\url{https://raw.githubusercontent.com/Hipo/university-domains-list/master/world_universities_and_domains.json}
}
\examples{
\dontrun{
uni_cat(domains = "http://www.google.com")
}
}
