% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virustotal_cat.R
\name{virustotal_cat}
\alias{virustotal_cat}
\title{Get Category from VirusTotal}
\usage{
virustotal_cat(domains = NULL, apikey = NULL)
}
\arguments{
\item{domains}{domain names as character vector}

\item{apikey}{virustotal API key}
}
\value{
data.frame with domain and VirusTotal analysis results
}
\description{
Returns category of content from multiple security vendors using the VirusTotal API v3.
The function retrieves domain analysis results including categories from various security
services. Not all services will have categories for all domains.
}
\details{
Get the API Access Key from \url{https://www.virustotal.com/}. Either pass the API Key to the function
or set the environmental variable: \code{VirustotalToken}. Environment variables persist within
a R session.
}
\examples{
\dontrun{
virustotal_cat("http://www.google.com")
virustotal_cat(c("google.com", "facebook.com"))
}
}
\references{
\url{https://docs.virustotal.com/reference/domains}
}
