% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virustotal_cat.R
\name{virustotal_cat}
\alias{virustotal_cat}
\title{Get Virustotal Category}
\usage{
virustotal_cat(domain = NULL, apikey = NULL)
}
\arguments{
\item{domain}{domain name}

\item{apikey}{virustotal API key}
}
\value{
data frame
}
\description{
Get Virustotal Category
}
\examples{
\dontrun{
virustotal_cat(domain="http://www.google.com")
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}

